feat: TNL-11173 authoring API offered as v0, not v1
