+++ cms/djangoapps/contentstore/rest_api/serializers/__init__.py
@@ -1,0 +1,4 @@
+"""
+Serializers for all contentstore API versions
+"""
+from .common import StrictSerializer
