+++ cms/djangoapps/contentstore/rest_api/serializers/common.py
