+++ cms/djangoapps/contentstore/rest_api/v0/__init__.py
@@ -1,0 +1,12 @@
+"""
+Views for v0 contentstore API.
+"""
+
+from cms.djangoapps.contentstore.rest_api.v0.views.assets import (
+    AssetsCreateRetrieveView,
+    AssetsUpdateDestroyView
+)
+from cms.djangoapps.contentstore.rest_api.v0.views.xblock import (
+    XblockView,
+    XblockCreateView
+)
