+++ cms/djangoapps/contentstore/rest_api/v0/serializers/__init__.py
@@ -2,4 +2,7 @@
 Serializers for v0 contentstore API.
 """
 from .advanced_settings import AdvancedSettingsFieldSerializer, CourseAdvancedSettingsSerializer
+from .assets import AssetSerializer
 from .tabs import CourseTabSerializer, CourseTabUpdateSerializer, TabIDLocatorSerializer
+from .transcripts import TranscriptSerializer
+from .xblock import XblockSerializer
