+++ cms/djangoapps/contentstore/rest_api/v0/serializers/xblock.py
@@ -2,7 +2,7 @@
 API Serializers for xblocks
 """
 from rest_framework import serializers
-from .common import StrictSerializer
+from cms.djangoapps.contentstore.rest_api.serializers.common import StrictSerializer
 
 # The XblockSerializer is designed to be scalable and generic. As such, its structure
 # should remain as general as possible. Avoid indiscriminately adding fields to it,
