+++ cms/djangoapps/contentstore/rest_api/v1/serializers/videos.py
@@ -2,7 +2,7 @@
 API Serializers for videos
 """
 from rest_framework import serializers
-from .common import StrictSerializer
+from cms.djangoapps.contentstore.rest_api.serializers.common import StrictSerializer
 
 
 class FileSpecSerializer(StrictSerializer):
