+++ cms/djangoapps/contentstore/rest_api/v1/views/__init__.py
@@ -8,15 +8,8 @@ from .grading import CourseGradingView
 from .proctoring import ProctoredExamSettingsView, ProctoringErrorsView
 from .home import HomePageView
 from .settings import CourseSettingsView
-from .xblock import XblockView, XblockCreateView
-from .assets import AssetsCreateRetrieveView, AssetsUpdateDestroyView
 from .videos import (
     CourseVideosView,
-    VideosUploadsView,
-    VideosCreateUploadView,
-    VideoImagesView,
-    VideoEncodingsDownloadView,
-    VideoFeaturesView,
     VideoUsageView,
 )
 from .help_urls import HelpUrlsView
